moved error message shown if skill level thresholds in test are missing from tab info to tab test
